Make `Color` `constexpr`

+++ src/include/imgui/imconfig.h
@@ -16,7 +16,7 @@
 
 #include <cstdint>
 
-#define ImTextureID std::uint64_t
+#define ImTextureID std::uintptr_t
 
 //---- Define assertion handler. Defaults to calling assert().
 // If your macro uses multiple statements, make sure is enclosed in a 'do { .. } while (0)' block so it can be used as a single statement.
